backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.23.3
httpx==0.27.2
websockets>=15.0
PyJWT>=2.10.1
cryptography>=41.0.7
google-generativeai>=0.3.0

